# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Rucodel Production Planning API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /driver/requests:
    get:
      summary: List the authenticated driver's requests
      responses:
        "200":
          description: Driver request page
    post:
      summary: Create a work-intake request as a driver
      requestBody:
        required: true
        content:
          application/json:
            example:
              customerId: "00000000-0000-0000-0000-000000000001"
              wheelQuantities:
                - type: BIPARTITE
                  quantity: 4
                - type: WASHED
                  quantity: 0
                - type: NORMAL
                  quantity: 15
              expectedWheelQuantity: 19 # deprecated compatibility field, treated as NORMAL only when wheelQuantities is absent
              expectedFactoryDropOffWindowStart: "2026-09-02T13:00:00+01:00"
              expectedFactoryDropOffWindowEnd: "2026-09-02T14:00:00+01:00"
              requestedFactoryPickupWindowStart: "2026-09-04T13:00:00+01:00"
              requestedFactoryPickupWindowEnd: "2026-09-04T14:00:00+01:00"
              notes: "Cliente pediu prioridade se possível"
      responses:
        "201":
          description: Created request
  /driver/requests/{requestId}/arrival:
    post:
      summary: Confirm physical factory arrival for the authenticated driver's own request
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Request moved to AT_FACTORY
...
  /admin/requests/{requestId}/arrival:
    post:
      summary: Confirm actual factory arrival and optionally received quantity
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated request
...
  /admin/factory-arrivals:
    get:
      summary: List communicated requests waiting for physical factory arrival confirmation
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [COMMUNICATED]
      responses:
        "200":
          description: Requests waiting for arrival confirmation
  /admin/factory-arrivals/{requestId}:
    get:
      summary: Retrieve a request in the factory-arrival workflow
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Request details
  /admin/factory-arrivals/{requestId}/confirm:
    post:
      summary: Confirm physical arrival of communicated wheels at the factory
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            example:
              actualArrivalAt: "2026-09-26T10:35:00+01:00"
      responses:
        "200":
          description: Request moved to AT_FACTORY
...
  /admin/drivers/{id}:
    get:
      summary: Administrator driver detail with linked messaging identities
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Driver detail
    patch:
      summary: Update administrator-managed driver data
      responses:
        "200":
          description: Updated driver
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer